move preview modal to another file
