+++ sdks/second-stage-form/src/services/custom/OnUploadFiles.ts
@@ -1,5 +1,6 @@
 import type { Model, UploadFilesEvent } from 'survey-core'
 import { ICustomOnUploadFiles, SurveyModelCustomizer, SurveyFileUploadValidatorScheme } from '../../interfaces'
+import { getPreviewModalHTML } from './ComponentUtils'
 
 export class CustomOnUploadFilesBuilder implements ICustomOnUploadFiles {
   public static make(): CustomOnUploadFilesBuilder {
@@ -56,85 +57,13 @@ export class CustomOnUploadFilesBuilder implements ICustomOnUploadFiles {
     const customProperties = jsonObj.custom as SurveyFileUploadValidatorScheme
     if (!customProperties.enableZoomInModal) return
 
-    const HTML_CONTENT = `
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    </head>
-    <body>
-      <div id="preview-modal" class="modal">
-        <span class="close close-button">&times;</span>
-        <img class="modal-content" id="image-modal">
-      </div>
-    
-      <script>
-        // Get the <span> element that closes the modal
-        var closedButton = document.getElementsByClassName("close-button")[0]
-    
-        // When the user clicks on <span> (x), close the modal
-        closedButton.onclick = function() {
-          modal.style.display = "none"
-        }
-      </script>
-    
-      <style>
-        .modal {
-          display: none; 
-          position: fixed; 
-          z-index: 2; 
-          padding-top: 100px; 
-          padding-bottom: 100px;
-          left: 0;
-          top: 0;
-          width: 100%; 
-          height: 100%;
-          overflow: auto; 
-          background-color: rgb(0,0,0); 
-          background-color: rgba(0,0,0,0.7);
-        }
-    
-        .modal-content {
-          margin: auto;
-          display: block;
-          width: 80%;
-          max-width: ${customProperties.modalMaxWidth || 700}px;
-        }
-    
-        .close {
-          position: absolute;
-          top: 15px;
-          right: 35px;
-          color: #f1f1f1;
-          font-size: 25px;
-          font-weight: bold;
-          transition: 0.3s;
-        }
-    
-        .close:hover,
-        .close:focus {
-          color: #bbb;
-          text-decoration: none;
-          cursor: pointer;
-        }
-    
-        @media only screen and (max-width: 700px){
-          .modal-content {
-            width: 100%;
-          }
-        }
-      </style>
-    
-    </body>
-    </html>    
-    `
+    const html = getPreviewModalHTML(customProperties)
 
     let modal = document.body.querySelector('.modal') as HTMLElement
     if (!modal) {
       const resp = await fetch('preview-modal.html')
-      const html = await resp.text()
-      document.body.insertAdjacentHTML('beforeend', HTML_CONTENT)
+      // const html = await resp.text()
+      document.body.insertAdjacentHTML('beforeend', html)
       modal = document.body.querySelector('.modal') as HTMLElement
     }
